pyproject.toml
[project]
name = "python-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pyinstaller>=6.16.0",
    "python-dotenv>=1.2.1",
    "selenium>=4.38.0",
]
